Add cleanup message

+++ cleanUp.sh
@@ -6,5 +6,11 @@ BAM=$1
 
 mkdir -p cache
 
-mv $BAM ${BAM/.bam/.bai} cache
+if [ -e ${BAM/.bam/_FixHdr.bai} ]; then
+    echo "Moving" $BAM
+    mv $BAM ${BAM/.bam/.bai} cache
+else
+    echo "FixHdr.bai does not exist, do not delete"
+    echo "Rerun fixHeader" $BAM
+fi
